+++ live-learning/src/main/java/ca/tango/livelearning/domain/StudentAnswer.java
@@ -9,7 +9,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class StudentAnswer<T> {
     @Id
     String id;
-    Integer questionId;
+    Long questionId;
     String studentId;
     T answer;
     Boolean isCorrectAnswer;
