+++ live-learning/src/main/java/ca/tango/livelearning/repository/AnswerBreakpointRepository.java
@@ -3,5 +3,5 @@ package ca.tango.livelearning.repository;
 import ca.tango.livelearning.domain.AnswerBreakpoint;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface AnswerBreakpointRepository extends MongoRepository<AnswerBreakpoint, Integer> {
+public interface AnswerBreakpointRepository extends MongoRepository<AnswerBreakpoint, Long> {
 }
