+++ live-learning/src/main/java/ca/tango/livelearning/repository/BreakpointRepository.java
@@ -3,5 +3,5 @@ package ca.tango.livelearning.repository;
 import ca.tango.livelearning.domain.Breakpoint;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface BreakpointRepository extends MongoRepository<Breakpoint, Integer> {
+public interface BreakpointRepository extends MongoRepository<Breakpoint, Long> {
 }
